requirements/local.txt
asgiref==3.2.7
autopep8==1.5.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
defusedxml==0.6.0
Django==3.0.5
django-extensions==2.2.9
django-secure==1.0.2
django-sslserver==0.22
idna==2.9
oauthlib==3.1.0
Pillow==7.1.2
psycopg2==2.8.5
pycodestyle==2.5.0
pycparser==2.20
PyJWT==1.7.1
pyOpenSSL==19.1.0
python3-openid==3.1.0
pytz==2019.3
requests==2.23.0
requests-oauthlib==1.3.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.3
sorl-thumbnail==12.6.3
sqlparse==0.3.1
urllib3==1.25.9
Werkzeug==1.0.1
